refactor: substitui parseInt por parseFloat para não ter que especificar radix

+++ src/App.js
@@ -75,7 +75,7 @@ class App extends Component {
 
     const attrValues = [cardAttr1, cardAttr2, cardAttr3];
     const checkAttrValues = attrValues.every((attr) => attr >= MIN && attr <= MAX);
-    const sumAttr = attrValues.reduce((soma, i) => (parseInt(soma) + parseInt(i)));
+    const sumAttr = attrValues.reduce((soma, i) => (parseFloat(soma) + parseFloat(i)));
 
     const isValid = emptyFields && checkAttrValues && sumAttr <= SUM;
     if (isValid) {
